gore-detection-system/requirements-windows-directml.txt
tensorflow-cpu==2.10
tensorflow-directml-plugin
numpy<2
scikit-learn
pillow
